requirements.txt
numpy
pandas
alpaca-trade-api
requests